--- base_ws/src/vision/src/yolov7_ros/src/detect_sign.py
@@ -1,192 +1,0 @@
-#!/usr/bin/python3
-
-
-from models.experimental import attempt_load
-from utils.general import non_max_suppression
-from utils.ros import create_detection_msg
-from utils.visualizer import draw_detections
-from ultralytics import YOLO
-import os
-from typing import Tuple, Union, List
-
-import torch
-import cv2
-from torchvision.transforms import ToTensor
-import numpy as np
-import rospy
-
-from std_msgs.msg import Header
-from vision_msgs.msg import Detection2DArray, Detection2D, BoundingBox2D
-from sensor_msgs.msg import Image
-from cv_bridge import CvBridge
-
-import sys
-
-
-def parse_classes_file(path):
-    classes = []
-    with open(path, "r") as f:
-        for line in f:
-            line = line.replace("\n", "")
-            classes.append(line)
-    return classes
-
-
-
-
-class Yolov11:
-    def __init__(self, weights, conf_thresh: float = 0.5, iou_thresh: float = 0.45,
-                 device: str = "cuda"):
-        self.conf_thresh = conf_thresh
-        self.iou_thresh = iou_thresh
-        self.device = device
-        self.model = attempt_load(weights, map_location=device)
-        self.model.eval()
-
-    @torch.no_grad()
-    def inference(self, img: torch.Tensor):
-        """
-        :param img: tensor [c, h, w]
-        :returns: tensor of shape [num_boxes, 6], where each item is represented as
-            [x1, y1, x2, y2, confidence, class_id]
-        """
-        img = img.unsqueeze(0)
-        pred_results = self.model(img)[0]
-        detections = non_max_suppression(
-            pred_results, conf_thres=self.conf_thresh, iou_thres=self.iou_thresh
-        )
-        if detections:
-            detections = detections[0]
-        return detections
-
-
-class Yolov11Publisher:
-    def __init__(self, img_topic: str, weights: str, conf_thresh: float = 0.5,
-                 iou_thresh: float = 0.45, pub_topic: str = "yolov7_detections",
-                 device: str = "cuda",
-                 img_size: Union[Tuple[int, int], None] = (640, 640),
-                 queue_size: int = 1, visualize: bool = False,
-                 class_labels: Union[List, None] = None):
-        """
-        :param img_topic: name of the image topic to listen to
-        :param weights: path/to/yolo_weights.pt
-        :param conf_thresh: confidence threshold
-        :param iou_thresh: intersection over union threshold
-        :param pub_topic: name of the output topic (will be published under the
-            namespace '/yolov7')
-        :param device: device to do inference on (e.g., 'cuda' or 'cpu')
-        :param queue_size: queue size for publishers
-        :visualize: flag to enable publishing the detections visualized in the image
-        :param img_size: (height, width) to which the img is resized before being
-            fed into the yolo network. Final output coordinates will be rescaled to
-            the original img size.
-        :param class_labels: List of length num_classes, containing the class
-            labels. The i-th element in this list corresponds to the i-th
-            class id. Only for viszalization. If it is None, then no class
-            labels are visualized.
-        """
-        self.img_size = img_size
-        self.device = device
-        self.class_labels = class_labels
-
-        vis_topic = pub_topic + "visualization" if pub_topic.endswith("/") else \
-            pub_topic + "/visualization"
-        self.visualization_publisher = rospy.Publisher(
-            vis_topic, Image, queue_size=queue_size
-        ) if visualize else None
-
-        self.bridge = CvBridge()
-
-        self.tensorize = ToTensor()
-        self.model = YOLO(weights)
-
-        self.img_subscriber = rospy.Subscriber(
-            img_topic, Image, self.process_img_msg
-        )
-        self.detection_publisher = rospy.Publisher(
-            pub_topic, Detection2DArray, queue_size=queue_size
-        )
-        rospy.loginfo("YOLOv11 initialization complete. Ready to start inference")
-
-    def process_img_msg(self, img_msg: Image):
-        """ callback function for publisher """
-        cv_image = self.bridge.imgmsg_to_cv2(img_msg)              
-
-        detections =self.model.predict(
-                source =cv_image,
-                classes =[ i for i in range(1,26)],
-                verbose = False,
-        )
-
-        detection_msg = create_detection_msg(img_msg, detections)
-        # rospy.loginfo(detection_msg)
-        self.detection_publisher.publish(detection_msg)
-
-        # visualizing if required
-        if self.visualization_publisher:
-            bboxes = [[int(box.xyxy[0][0]), int(box.xyxy[0][1]), int(box.xyxy[0][2]), int(box.xyxy[0][3])]for box in detections[0].boxes]
-            classes = [int(box.cls) for  box in detections[0].boxes]
-            cv_image = draw_detections(cv_image, bboxes, classes,
-                                      self.class_labels)
-            # classes = [int(c) for c in detections[:, 5].tolist()]
-            self.publish_image(cv_image, cv_image.shape[0], cv_image.shape[1])
-
-
-    def publish_image(self, imagedata, height, width):
-        image_temp = Image()
-        header = Header(stamp=rospy.Time.now())
-        image_temp.height = height
-        image_temp.width = width
-        image_temp.encoding = 'bgr8'
-        image_temp.data = np.array(imagedata).tobytes()
-        image_temp.header = header
-        image_temp.step = width * 3
-        self.visualization_publisher.publish(image_temp)
-
-
-if __name__ == "__main__":
-    rospy.init_node("yolov11_node")
-
-    ns = rospy.get_name() + "/"
-
-    weights_path = rospy.get_param(ns + "weights_path")
-    classes_path = rospy.get_param(ns + "classes_path")
-    img_topic = rospy.get_param(ns + "img_topic")
-    out_topic = rospy.get_param(ns + "out_topic")
-    conf_thresh = rospy.get_param(ns + "conf_thresh")
-    iou_thresh = rospy.get_param(ns + "iou_thresh")
-    queue_size = rospy.get_param(ns + "queue_size")
-    img_size = rospy.get_param(ns + "img_size")
-    visualize = rospy.get_param(ns + "visualize")
-    device = rospy.get_param(ns + "device")
-
-    # some sanity checks
-    if not os.path.isfile(weights_path):
-        raise FileExistsError(f"Weights not found ({weights_path}).")
-    
-    if classes_path: 
-        if not os.path.isfile(classes_path):
-            raise FileExistsError(f"Classes file not found ({classes_path}).")
-        classes = parse_classes_file(classes_path)
-    else:
-        rospy.loginfo("No class file provided. Class labels will not be visualized.")
-        classes = None
-
-    if not ("cuda" in device or "cpu" in device):
-        raise ValueError("Check your device.")
-
-
-    publisher = Yolov11Publisher(
-        img_topic=img_topic,
-        pub_topic=out_topic,
-        weights=weights_path,
-        device=device,
-        visualize=visualize,
-        conf_thresh=conf_thresh,
-        iou_thresh=iou_thresh,
-        img_size=(img_size, img_size),
-        queue_size=queue_size,
-        class_labels=classes
-    )
-
-    rospy.spin()
