--- base_ws/src/vision/src/yolov7_ros/src/train_yolov7.sh
@@ -1,4 +1,0 @@
-python3 train.py  \
---weights 'yolo7_weights/best_0911.pt' \
---data    'data/data_20241030/data.yaml' \
---device 0 \
